Resolved PR Comments

+++ solar_offset/views/admin.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, session, redirect
+from flask import Blueprint, render_template, request, redirect, g
 from solar_offset.db import get_db
 from solar_offset.views.auth import login_required
 
@@ -9,16 +9,15 @@ bp = Blueprint("admin", __name__)
 @login_required("a")
 def admin():
     db = get_db()
-    user_types = ["admin", "householder", "staff"]
+    current_admin = g.user['id']
     users = db.execute(
-        'SELECT * FROM user WHERE user_type NOT LIKE ? ', ('%a%',)
+        'SELECT * FROM user WHERE id!= ? ', (current_admin,)
     ).fetchall()
     user_status_list = db.execute('SELECT * FROM user_status ').fetchall()
 
-    user_status_dict = []
+    user_status_dict = {}
     for user_row in user_status_list:
-        userstatusdict = dict(user_row)
-        user_status_dict.append(userstatusdict)
+        user_status_dict[user_row['user_id']] = user_row['suspend']
 
     user_dicts = []
     for user_row in users:
@@ -27,13 +26,13 @@ def admin():
             userdict["user_type"] = "householder"
         else:
             userdict["user_type"] = "staff"
+        userdict["is_suspended"] = user_status_dict.get(user_row['id'], "Null")
         user_dicts.append(userdict)
 
-        user_statuses = [item['user_id'] for item in user_status_dict]
     return render_template(
         "./users/admin/admin.html",
         users=user_dicts,
-        user_status_list=user_statuses)
+    )
 
 
 @bp.route('/delete_user', methods=['POST'])
@@ -45,24 +44,30 @@ def delete_user():
     return redirect('/admin')
 
 
-@bp.route('/suspend-user', methods=['POST'])
-def suspend_user():
-    suspend_message = request.form['suspend_message']
+@bp.route('/is-suspend-user', methods=['POST'])
+def is_suspend_user():
     user_id = request.form['user_id']
     db = get_db()
-    db.execute(
-        "INSERT INTO user_status (user_id, suspend) VALUES (?,?)", (user_id, suspend_message,)
-    )
+    if 'suspend_message' in request.form:
+        suspend_message = request.form['suspend_message']
+
+        db.execute(
+            "INSERT INTO user_status (user_id, suspend) VALUES (?,?)", (user_id, suspend_message,)
+        )
+    else:
+        db.execute(
+            "DELETE FROM user_status WHERE user_id = ?", (user_id,)
+        ).fetchone()
+
     db.commit()
     return redirect('/admin')
 
-
-@bp.route('/unsuspend-user', methods=['POST'])
-def unsuspend_user():
-    user_id = request.form['user_id']
-    db = get_db()
-    db.execute(
-        "DELETE FROM user_status WHERE user_id = ?", (user_id,)
-    ).fetchone()
-    db.commit()
-    return redirect('/admin')
+# @bp.route('/unsuspend-user', methods=['POST'])
+# def unsuspend_user():
+#     user_id = request.form['user_id']
+#     db = get_db()
+#     db.execute(
+#         "DELETE FROM user_status WHERE user_id = ?", (user_id,)
+#     ).fetchone()
+#     db.commit()
+#     return redirect('/admin')
